+++ app/gallery/eventData.js
@@ -1,56 +1,56 @@
-export const upcomingEvents = [
-    {
-        title: "Fortune 404",
-        description: "Get ready for an adrenaline-charged Treasure Hunt that takes you across campus to crack clever riddles, race the clock, and unlock your starting stash. Then dive into the Investment Arena, where every twist in the storyline challenges your wits: decode clues, outsmart puzzles, and choose between bold, high-stakes moves or steady, calculated plays.",
-        date: "19th September 2025",
-        time: "10:00 AM",
-        location: "MEL Lab 03, PJ Block, 2nd Floor",
-        category: "Phaseshift 2025",
-        status: "completed",
-        registrationFee: "₹100 per team",
-        prizePool: "₹5000",
-        coordinators: [
-            { name: "Shriyans Naik", number: "+91 79825 31455" },
-            { name: "Divyam Jain", number: "+91 60099 13709" }
-        ],
-        registrationLink: "https://events.phaseshift.in/events/GRDFFO",
-        poster: "https://gradient-content-server.vercel.app/content/ps25/f404poster.png"
-    },
-    {
-        title: "AI Unplugged",
-        description: "Get set for an immersive Hands-On Workshop that unpacks the secrets of RAG (Retrieval-Augmented Generation) and Prompt Engineering. Guided by Tulsi Patro, Prompt Engineer at Amazon, you’ll learn how to craft precise prompts, and see how RAG supercharges LLMs. Whether you’re new or sharpening your skills, It will be an innovative experience.",
-        date: "20th September 2025",
-        time: "2:00 PM",
-        location: "MEL-Seminar Hall, 7th floor",
-        category: "Phaseshift 2025",
-        status: "completed",
-        registrationFee: "₹50 Per Person",
-        prizePool: "N/A",
-        coordinators: [
-            { name: "Channabasavanna MB", number: "+91 76769 75495" },
-            { name: "Riya Kumari", number: "+91 82521 89462" }
-        ],
-        registrationLink: "https://events.phaseshift.in/events/GRDAUD",
-        poster: "https://gradient-content-server.vercel.app/content/ps25/aiunplugposter.png"
-    },
-    // {
-    //     title: "Sync or Sink",
-    //     description: "A test of teamwork and strategy, where participants must work together to solve puzzles and challenges. You either Sync up with your teammates or Sink into the depths of despair.",
-    //     date: "May 10th, 2025",
-    //     time: "2:30 PM",
-    //     location: "CSE Seminar Hall, 4th Floor, PJA Block, BMSCE",
-    //     category: "Utsav 2025",
-    //     status: "registration-open",
-    //     registrationFee: "₹100 per team",
-    //     prizePool: "₹4,000",
-    //     coordinators: [
-    //         { name: "Yashas Nandan", number: "+91 7676823822" },
-    //         { name: "Vignesh Madan", number: "+91 7022433357" }
-    //     ],
-    //     registrationLink: "https://events.bmsutsav.in/events/GRDSOS",
-    //     poster: "https://gradient-content-server.vercel.app/content/utsav25/SOS.png"
-    // }
-];
+// export const upcomingEvents = [
+//     {
+//         title: "Fortune 404",
+//         description: "Get ready for an adrenaline-charged Treasure Hunt that takes you across campus to crack clever riddles, race the clock, and unlock your starting stash. Then dive into the Investment Arena, where every twist in the storyline challenges your wits: decode clues, outsmart puzzles, and choose between bold, high-stakes moves or steady, calculated plays.",
+//         date: "19th September 2025",
+//         time: "10:00 AM",
+//         location: "MEL Lab 03, PJ Block, 2nd Floor",
+//         category: "Phaseshift 2025",
+//         status: "completed",
+//         registrationFee: "₹100 per team",
+//         prizePool: "₹5000",
+//         coordinators: [
+//             { name: "Shriyans Naik", number: "+91 79825 31455" },
+//             { name: "Divyam Jain", number: "+91 60099 13709" }
+//         ],
+//         registrationLink: "https://events.phaseshift.in/events/GRDFFO",
+//         poster: "https://gradient-content-server.vercel.app/content/ps25/f404poster.png"
+//     },
+//     {
+//         title: "AI Unplugged",
+//         description: "Get set for an immersive Hands-On Workshop that unpacks the secrets of RAG (Retrieval-Augmented Generation) and Prompt Engineering. Guided by Tulsi Patro, Prompt Engineer at Amazon, you’ll learn how to craft precise prompts, and see how RAG supercharges LLMs. Whether you’re new or sharpening your skills, It will be an innovative experience.",
+//         date: "20th September 2025",
+//         time: "2:00 PM",
+//         location: "MEL-Seminar Hall, 7th floor",
+//         category: "Phaseshift 2025",
+//         status: "completed",
+//         registrationFee: "₹50 Per Person",
+//         prizePool: "N/A",
+//         coordinators: [
+//             { name: "Channabasavanna MB", number: "+91 76769 75495" },
+//             { name: "Riya Kumari", number: "+91 82521 89462" }
+//         ],
+//         registrationLink: "https://events.phaseshift.in/events/GRDAUD",
+//         poster: "https://gradient-content-server.vercel.app/content/ps25/aiunplugposter.png"
+//     },
+// {
+//     title: "Sync or Sink",
+//     description: "A test of teamwork and strategy, where participants must work together to solve puzzles and challenges. You either Sync up with your teammates or Sink into the depths of despair.",
+//     date: "May 10th, 2025",
+//     time: "2:30 PM",
+//     location: "CSE Seminar Hall, 4th Floor, PJA Block, BMSCE",
+//     category: "Utsav 2025",
+//     status: "registration-open",
+//     registrationFee: "₹100 per team",
+//     prizePool: "₹4,000",
+//     coordinators: [
+//         { name: "Yashas Nandan", number: "+91 7676823822" },
+//         { name: "Vignesh Madan", number: "+91 7022433357" }
+//     ],
+//     registrationLink: "https://events.bmsutsav.in/events/GRDSOS",
+//     poster: "https://gradient-content-server.vercel.app/content/utsav25/SOS.png"
+// }
+// ];
 export const eventCategories = [
     {
         name: "Utsav",
@@ -137,6 +137,100 @@ export const eventCategories = [
                         description: "An event featuring time-themed puzzles, decoding ancient ciphers, and trivia with unique twists, combining education and competition.",
                         status: "completed",
                         poster: "https://gradient-content-server.vercel.app/content/posters/tmr.webp"
+                    }
+                ]
+            }
+        ]
+    },
+    {
+        name: "Phaseshift",
+        description: "The flagship technical symposium of BMSCE where Gradient organized various competitions, workshops, and tech showcases.",
+        years: [
+            {
+                year: "2025",
+                events: [
+                    {
+                        title: "Fortune 404",
+                        participants: 60,
+                        description: "An adrenaline-charged Treasure Hunt that took participants across campus to crack clever riddles, race the clock, and unlock their starting stash. Then dive into the Investment Arena, where every twist in the storyline challenged wits: decode clues, outsmart puzzles, and choose between bold, high-stakes moves or steady, calculated plays.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/ps25/f404poster.png",
+                        images: [
+                            "fffs.png",                            
+                            "https://gradient-content-server.vercel.app/content/ps25/1.webp",
+                            "https://gradient-content-server.vercel.app/content/ps25/3.webp",
+                            "https://gradient-content-server.vercel.app/content/ps25/4.webp"
+                        ]
+                    },
+                    {
+                        title: "AI Unplugged",
+                        participants: 50,
+                        description: "An immersive Hands-On Workshop that unpacked the secrets of RAG (Retrieval-Augmented Generation) and Prompt Engineering. Guided by Tulsi Patro, Prompt Engineer at Amazon, participants learned how to craft precise prompts, and saw how RAG supercharges LLMs.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/ps25/aiunplugposter.png",
+                        images: [
+                            "https://gradient-content-server.vercel.app/content/ps25/5.webp",
+                            "https://gradient-content-server.vercel.app/content/ps25/6.webp",
+                        ]
+                    }
+                ]
+            },
+            {
+                year: "2024",
+                events: [
+                    {
+                        title: "Maze Maniac Resurrected",
+                        participants: 56,
+                        description: "A solo Minecraft-based event featuring tricky mazes, parkour runs, PvE battles, and clever puzzles. Participants relied on their skills to navigate challenges, creating an unforgettable experience.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/mmrlogo.webp"
+                    },
+                    {
+                        title: "Cipher Secrets",
+                        participants: 38,
+                        description: "An interactive event featuring diverse puzzles and cryptographic tasks, guided by a chatbot that provided hints and instructions. It combined problem-solving and teamwork in a high-energy adventure.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/ciphersec.webp"
+                    },
+                    {
+                        title: "Retro Design 2.0",
+                        participants: 26,
+                        description: "A reverse engineering challenge testing creativity and problem-solving abilities with tasks like uncovering hidden clues and solving intricate mysteries. ₹4,000 prize pool.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/rd2.webp"
+                    }
+                ]
+            },
+            {
+                year: "2023",
+                events: [
+                    {
+                        title: "Maze Maniac 2.0",
+                        participants: 40,
+                        description: "A thrilling Minecraft event where participants navigated increasingly complex mazes using strategy and creativity. The event showcased the excitement and tactical genius of Minecraft.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/mmr2.webp"
+                    },
+                    {
+                        title: "Retro Design",
+                        participants: 29,
+                        description: "A reverse engineering event where participants solved executable files to uncover encrypted messages, using crossword puzzles to decrypt the final message.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/rd1.webp"
+                    },
+                    {
+                        title: "Datathon Hackathon",
+                        participants: 11,
+                        description: "An AI hackathon testing skills in data preprocessing and model building with CNNs and ANNs. Participants developed models without pre-trained datasets and solved custom test cases.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/datathon.webp"
+                    },
+                    {
+                        title: "Workshop on LLM and LangChain",
+                        participants: 26,
+                        description: "A hands-on workshop conducted by Mr. Ravi Ranjan, introducing students to LLMs and LangChain, with engaging sessions and practical applications.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/posters/pswork1.webp"
                     }
                 ]
             }
@@ -252,74 +346,28 @@ export const eventCategories = [
         ]
     },
     {
-        name: "Phaseshift",
-        description: "The flagship technical symposium of BMSCE where Gradient organized various competitions, workshops, and tech showcases.",
-        years: [
-            {
-                year: "2024",
-                events: [
-                    {
-                        title: "Maze Maniac Resurrected",
-                        participants: 56,
-                        description: "A solo Minecraft-based event featuring tricky mazes, parkour runs, PvE battles, and clever puzzles. Participants relied on their skills to navigate challenges, creating an unforgettable experience.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/mmrlogo.webp"
-                    },
-                    {
-                        title: "Cipher Secrets",
-                        participants: 38,
-                        description: "An interactive event featuring diverse puzzles and cryptographic tasks, guided by a chatbot that provided hints and instructions. It combined problem-solving and teamwork in a high-energy adventure.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/ciphersec.webp"
-                    },
-                    {
-                        title: "Retro Design 2.0",
-                        participants: 26,
-                        description: "A reverse engineering challenge testing creativity and problem-solving abilities with tasks like uncovering hidden clues and solving intricate mysteries. ₹4,000 prize pool.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/rd2.webp"
-                    }
-                ]
-            },
-            {
-                year: "2023",
-                events: [
-                    {
-                        title: "Maze Maniac 2.0",
-                        participants: 40,
-                        description: "A thrilling Minecraft event where participants navigated increasingly complex mazes using strategy and creativity. The event showcased the excitement and tactical genius of Minecraft.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/mmr2.webp"
-                    },
-                    {
-                        title: "Retro Design",
-                        participants: 29,
-                        description: "A reverse engineering event where participants solved executable files to uncover encrypted messages, using crossword puzzles to decrypt the final message.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/rd1.webp"
-                    },
-                    {
-                        title: "Datathon Hackathon",
-                        participants: 11,
-                        description: "An AI hackathon testing skills in data preprocessing and model building with CNNs and ANNs. Participants developed models without pre-trained datasets and solved custom test cases.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/datathon.webp"
-                    },
-                    {
-                        title: "Workshop on LLM and LangChain",
-                        participants: 26,
-                        description: "A hands-on workshop conducted by Mr. Ravi Ranjan, introducing students to LLMs and LangChain, with engaging sessions and practical applications.",
-                        status: "completed",
-                        poster: "https://gradient-content-server.vercel.app/content/posters/pswork1.webp"
-                    }
-                ]
-            }
-        ]
-    },
-    {
         name: "Standalone",
         description: "Special events and one-off initiatives throughout the academic year to enrich the mind and fulfill the soul.",
         years: [
+            {
+                year: "2025",
+                events: [
+                    {
+                        title: "AI AMPED Workshop",
+                        participants: 70,
+                        description: "A hands-on workshop conducted by Mr. Shivam Modi. Particpants engaged in key concepts such as Gen AI and uncovering frameworks of Transformers while shedding light on the convergence of AI, Electronics and Semiconductors.",
+                        status: "completed",
+                        poster: "https://gradient-content-server.vercel.app/content/gallery/sa/2025/Ai amped.webp",
+                        images: [
+                            "https://media.licdn.com/dms/image/v2/D5622AQHhC7TtjlVWIA/feedshare-shrink_2048_1536/B56Zntc0D3J8Aw-/0/1760625381959?e=1763596800&v=beta&t=P3Fh8VnBeXOjpvitE-igJJ-lJaq3nTGJS5iP95XHMHA",
+                            "https://media.licdn.com/dms/image/v2/D5622AQGwzRsXbbKGVQ/feedshare-shrink_2048_1536/B56Zntc0DGKAAw-/0/1760625379353?e=1763596800&v=beta&t=BGz_8KI9L45xuTa9xVV3fnWfuhPIqlGKfIgFfP5iFvg",
+                            "https://media.licdn.com/dms/image/v2/D5622AQEuip1pl00ocw/feedshare-shrink_2048_1536/B56Zntc0DBJsAw-/0/1760625378791?e=1763596800&v=beta&t=XZI8il2SGAH1VtuAZwqJqCRY46QOA3NYI1o3yaaKrco",
+                            "https://media.licdn.com/dms/image/v2/D5622AQEk50MIJ987cQ/feedshare-shrink_2048_1536/B56Zntc0CyJsAw-/0/1760625378421?e=1763596800&v=beta&t=KrafgT0tt0-c1nLmX6n2Et7CNHBnBs0pbMJ0G0mdRB8"
+
+                        ]
+                    }
+                ]
+            },
             {
                 year: "2024",
                 events: [
@@ -383,5 +431,38 @@ export const eventCategories = [
                 ]
             }
         ]
+    }
+];
+
+export const specialShowcases = [
+    {
+        title: "Gradient Week 2025",
+        description: "Explore the official page for Gradient Week 2025, our flagship annual Tech Innovation Week featuring hackathons, workshops, and tech showcases.",
+        image: "https://gradient-content-server.vercel.app/content/gallery/recap/special/3.webp",
+        link: "https://gradientaiml.tech/gw25"
+    },
+    {
+        title: "Gradient Week 2025 Recap",
+        description: "Relive the highlights and memorable moments from Gradient Week 2025 through our comprehensive event recap.",
+        image: "https://gradient-content-server.vercel.app/content/gallery/recap/special/4.webp",
+        link: "https://gradientaiml.tech/gw25recap"
+    },
+    {
+        title: "Gradient X Utsav 2025",
+        description: "Discover our exciting collaboration with Utsav Ananta 2025, BMSCE's annual cultural festival, featuring unique events and showcases.",
+        image: "https://gradient-content-server.vercel.app/content/utsav25/gallery/special/5.webp",
+        link: "https://gradientaiml.tech/utsav25"
+    },
+    {
+        title: "Gradient X Utsav 2025 Recap",
+        description: "Browse through the unforgettable moments and achievements from our partnership with Utsav Ananta 2025.",
+        image: "https://gradient-content-server.vercel.app/content/utsav25/gallery/special/6.webp",
+        link: "https://gradientaiml.tech/utsav25recap"
+    },
+    {
+        title: "Gradient X PhaseShift 2025",
+        description: "Visit our dedicated page for PhaseShift Meridian 2025, BMSCE's premier technical symposium, showcasing our innovative events and workshops.",
+        image: "https://gradient-content-server.vercel.app/content/special/7.webp",
+        link: "https://gradientaiml.tech/ps25"
     }
 ];